back: security vol1. (#36)
* chore: gitignore firebase key

* chore: spring security & firebase-admin

* feat: Unauthorized message

* feat: JWT utils

* feat: Auth UserDetails

* feat: AuthTokenFilter

* feat: Security configuration

* refactor: cleaned code

* refactor: role based endpoint protection

* feat: Role model

* feat: Role repository

* feat: JSON getter for roles

* refactor: userDetail to roles

* feat: Auth user by roles

* feat: auth service

* refactor: removed old printouts

* feat: added googleId field!
Still needs to get googleId

* Revert "merge: rollback from googleId"

This reverts commit bfa65d286ae67849e770612288edf9a451a8eb66, reversing
changes made to ab73796658903358d59fb1db23b4670a6bf10b42.

* feat: use googleUid

* feat: use googleUid

* refactor: helper methods for clean code

* feat: POST project protected

* feat: PUT project protected

* refactor: helper method to check project owner

* feat: delete project protected

* fix: deleting project never resulted 404

* feat: GET users for admins

* feat: POST user protected

* feat: GETers protected

* feat: PUT protection

* feat: DELETE protection

* document: code example Auth header

* build: validation with security

* documentation: removed code example

* fix: front up to date with main

* refactor: cleaned code, added TODOs for security

* feat: protect user part of projects

* feat: protect project application

* feat: admin bypass joining project

* build: security

* fix: partOfProjects was null

* build: fixed error

* feat: read firebase credentials from local file

* build: reading credentials from resource file

+++ back/src/main/java/com/experis/lagalt/controllers/UserController.java
@@ -2,6 +2,7 @@ package com.experis.lagalt.controllers;
 
 import com.experis.lagalt.models.Project;
 import com.experis.lagalt.models.User;
+import com.experis.lagalt.services.AuthService;
 import com.experis.lagalt.services.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -19,8 +20,14 @@ public class UserController {
     @Autowired
     private UserService userService;
 
+    @Autowired
+    private AuthService authService;
+
     @GetMapping
     public ResponseEntity<List<User>> getUsers() {
+        if (!authService.isAdmin()) {
+            return new ResponseEntity<>(null, HttpStatus.FORBIDDEN);
+        }
         List<User> users = userService.getAllUsers();
         HttpStatus status = HttpStatus.OK;
         return new ResponseEntity<>(users, status);
@@ -28,6 +35,9 @@ public class UserController {
 
     @PostMapping
     public ResponseEntity<User> createUser(@Valid @RequestBody User user) {
+        if (!authService.isLoggedUser(user)) {
+            return new ResponseEntity<>(null, HttpStatus.FORBIDDEN);
+        }
         User newUser = userService.saveUser(user);
         HttpStatus status = HttpStatus.CREATED;
         return new ResponseEntity<>(newUser, status);
@@ -35,18 +45,22 @@ public class UserController {
 
     @GetMapping(value = "/{id}")
     public ResponseEntity<User> getUser(@PathVariable long id) {
-        User user = userService.findUser(id);
-        HttpStatus status;
-        if (userService.userExists(id)) {
-            status = HttpStatus.OK;
-        } else {
-            status = HttpStatus.NOT_FOUND;
+        if (!userService.userExists(id)) {
+            return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
         }
+        if (!authService.isLoggedUser(id)) {
+            return new ResponseEntity<>(null, HttpStatus.FORBIDDEN);
+        }
+        User user = userService.findUser(id);
+        HttpStatus status = HttpStatus.OK;
         return new ResponseEntity<>(user, status);
     }
 
     @PutMapping(value = "/{id}")
     public ResponseEntity<User> updateUser(@PathVariable long id, @Valid @RequestBody User newUser) {
+        if (!authService.isLoggedUser(newUser)) {
+            return new ResponseEntity<>(null, HttpStatus.FORBIDDEN);
+        }
         User returnUser = new User();
         HttpStatus status;
         if (id != newUser.getId()) {
@@ -64,49 +78,52 @@ public class UserController {
 
     @DeleteMapping(value = "/{id}")
     public ResponseEntity<User> deleteUser(@PathVariable long id) {
-        HttpStatus status;
-        boolean userDeleted = userService.deleteUser(id);
-        if (userDeleted) {
-            status = HttpStatus.NO_CONTENT;
-        } else {
-            status = HttpStatus.NOT_FOUND;
+        if (!userService.userExists(id)) {
+            return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
         }
+        if (!authService.isLoggedUser(id)) {
+            return new ResponseEntity<>(null, HttpStatus.FORBIDDEN);
+        }
+        userService.deleteUser(id);
+        HttpStatus status = HttpStatus.NO_CONTENT;
         return new ResponseEntity<>(null, status);
     }
 
     @GetMapping(value = "/{id}/projects")
     public ResponseEntity<List<Project>> getUserProjects(@PathVariable long id) {
-        List<Project> projects = userService.getUserProjects(id);
-        HttpStatus status;
-        if (userService.userExists(id)) {
-            status = HttpStatus.OK;
-        } else {
-            status = HttpStatus.NOT_FOUND;
+        if (!userService.userExists(id)) {
+            return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
         }
+        if (!authService.isLoggedUser(id)) {
+            return new ResponseEntity<>(null, HttpStatus.FORBIDDEN);
+        }
+        List<Project> projects = userService.getUserProjects(id);
+        HttpStatus status = HttpStatus.OK;
         return new ResponseEntity<>(projects, status);
+    }
+
+    @GetMapping(value = "/googleid/{googleId}")
+    public ResponseEntity<User> getUser(@PathVariable String googleId) {
+        if (!userService.userExists(googleId)) {
+            return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
+        }
+        if (!authService.isLoggedUser(googleId)) {
+            return new ResponseEntity<>(null, HttpStatus.FORBIDDEN);
+        }
+        User user = userService.findUser(googleId);
+        HttpStatus status = HttpStatus.OK;
+        return new ResponseEntity<>(user, status);
     }
 
     @GetMapping(value = "/{userId}/projects/participant")
-    public ResponseEntity<List<Project>> getProjectsUserPartOf(@PathVariable long userId){
+    public ResponseEntity<List<Project>> getProjectsUserPartOf(@PathVariable long userId) {
+        if (!userService.userExists(userId)) {
+            return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
+        }
+        if (!authService.isLoggedUser(userId)) {
+            return new ResponseEntity<>(null, HttpStatus.FORBIDDEN);
+        }
         List<Project> projects = userService.getUserProjectsPartOf(userId);
-        HttpStatus status;
-        if (userService.userExists(userId)) {
-            status = HttpStatus.OK;
-        } else {
-            status = HttpStatus.NOT_FOUND;
-        }
-        return new ResponseEntity<>(projects, status);
-    }
-
-    @GetMapping(value = "/googleid/{googleid}")
-    public ResponseEntity<User> getUser(@PathVariable String googleid) {
-        User user = userService.findUser(googleid);
-        HttpStatus status;
-        if (userService.userExists(googleid)) {
-            status = HttpStatus.OK;
-        } else {
-            status = HttpStatus.NOT_FOUND;
-        }
-        return new ResponseEntity<>(user, status);
+        return new ResponseEntity<>(projects, HttpStatus.OK);
     }
 }
